+++ src/NKSolver/blockette.F90
@@ -189,9 +189,7 @@ contains
 
        do sps=1, nTimeIntervalsSpectral
           ! Update overset connectivity if necessary
-          if (oversetPresent .and. &
-               (oversetUpdateMode == updateFast .or. &
-               oversetUpdateMode == updateFull)) then
+          if (oversetPresent .and. oversetUpdateMode == updateFast) then
              call updateOversetConnectivity(1_intType, sps)
           end if
        end do
